Returning mels not pure audio
